Refactored, and abstracted the functions in setup, wrote more tests
