+++ .env_test
@@ -9,18 +9,18 @@ NODE_ENV=development
 # This environment variable is used to provide a unique random secret key for
 # for signing/verifying access tokens using JWT(jsonwebtokens).
 
-ACCESS_TOKEN_SECRET=af91d9228eeddec9c34acce52965cad36e367bdce6f1fc1bc378d40a0ad97e67
+ACCESS_TOKEN_SECRET=e64303eff97a9ad9531add6ff277c49f64323b28373dc3a5364f36941f404905
 
 
 # This environment variable is used to provide a unique random secret key for
 # for signing/verifying refresh tokens using JWT(jsonwebtokens).
 
-REFRESH_TOKEN_SECRET=4a9f15c20f299008ecefcb51cee65d17e45a504a72dd68cd2ff098f138d89a74
+REFRESH_TOKEN_SECRET=cb7612ba06493937ad42fc8b9a10ebc6947a921751afe2f0f69bc4e4e4d15908
 
 
 # This environment variable is used to provide connection string of the mongoDB
 # database for talawa-api to connect to.
-MONGO_DB_URL=
+MONGO_DB_URL=mongodb://testUrl-new/talawa-api
 
 
 # This environment variable is used to provide reCAPTCHA secret key for 
@@ -32,11 +32,11 @@ RECAPTCHA_SECRET_KEY=
 # This environment variable is used to provide the username for mailing service.
 # authentication purpose.
 
-MAIL_USERNAME=
+MAIL_USERNAME=banerjeeanubhav03@gmail.com
 
 # This environment variable is used to provide the password for mailing service.
 
-MAIL_PASSWORD=
+MAIL_PASSWORD=skxu vbjf hpoc fljb
 
 # For using custom SMTP server. (Ignore if you are using GMAIL as a Mail Transport)
 # true/false
@@ -68,7 +68,7 @@ LOG_PATH=./logs/transaction.log
 #
 # **NOTE** It is STRONGLY advised that you remove the email address once the initial installation and setup has been done
 
-LAST_RESORT_SUPERADMIN_EMAIL=
+LAST_RESORT_SUPERADMIN_EMAIL=banerjeeanubhav03@gmail.com
 
 # Logger Configuration
 COLORIZE_LOGS = false
@@ -76,8 +76,8 @@ LOG_LEVEL = info
 
 # Redis Cache Config
 
-REDIS_HOST=localhost
-REDIS_PORT=6379
+REDIS_HOST=test
+REDIS_PORT=6378
 REDIS_PASSWORD=
 
 # this environment variable is for setting the environment variable for Image Upload size
